add styling, filters, sorting, customizing

+++ blogger/views.py
@@ -12,10 +12,13 @@ from django.urls import reverse
 from django.db.models import F
 from django.views import generic, View
 from django.utils import timezone
+from dateutil.relativedelta import relativedelta
 from .models import *
 from .forms import *
 
 # Create your views here.
+
+PREVIEW_TEXT_CHAR = 300;
 
 def can_access(user, blogpost_id):
 	blogpost = get_object_or_404(BlogPost, pk=blogpost_id)
@@ -73,15 +76,59 @@ class SignupView(View):
 
 class BlogView(View):
 	def get(self, request, blog_title):
-		# context['blog_title'] = blog_title
 		blog = get_object_or_404(Blog, pk=blog_title)
 		blogposts = blog.blogpost_set.all().order_by('-pub_date')
 		context = {
 			'blog': blog,
-			'blogposts': blogposts
+			'blogposts': [],
+			'dates': [],
 		}
+		for blogpost in blogposts:
+			image_count = blogpost.imagedoc_set.all().count()
+			if image_count>0:
+				context['blogposts'].append((blogpost,blogpost.imagedoc_set.all()[image_count-1], blogpost.blogpost_text[:PREVIEW_TEXT_CHAR]))
+			else:
+				context['blogposts'].append((blogpost, '', blogpost.blogpost_text[:PREVIEW_TEXT_CHAR]))
+		if blogposts.count()>0:
+			oldest_blog_date = blogposts[blogposts.count()-1].pub_date
+			now = timezone.now()
+			while now>=oldest_blog_date or (now.month==oldest_blog_date.month and now.year==oldest_blog_date.year):
+				context['dates'].append((now.strftime('%B'),now.year))
+				now = now - relativedelta(months=1)
 		if self.request.user.is_authenticated():
 			context['username'] = request.user.username
+		return render(request, 'blogger/blog.html', context)
+
+	def post(self, request, blog_title):
+		blog = get_object_or_404(Blog, user=request.user)
+		context = {
+			'username': request.user.username,
+			'blog': blog,
+			'blogposts': [],
+			'dates': [],
+		}
+		if 'datefilter' in request.POST:
+			datetime = request.POST['datefilter']
+			if datetime == 'All Time':
+				blogposts = blog.blogpost_set.all().order_by('-pub_date')
+			else:
+				datetime = datetime.split(' ')
+				month = convertMonthNameToNumber(datetime[0])
+				year = datetime[1]
+				blogposts = blog.blogpost_set.all().order_by('-pub_date').filter(pub_date__month=month, pub_date__year=year)
+		for blogpost in blogposts:
+			image_count = blogpost.imagedoc_set.all().count()
+			if image_count>0:
+				context['blogposts'].append((blogpost,blogpost.imagedoc_set.all()[image_count-1], blogpost.blogpost_text[:PREVIEW_TEXT_CHAR]))
+			else:
+				context['blogposts'].append((blogpost, '', blogpost.blogpost_text[:PREVIEW_TEXT_CHAR]))
+		blogposts = blog.blogpost_set.all().order_by('-pub_date')
+		if blogposts.count()>0:
+			oldest_blog_date = blogposts[blogposts.count()-1].pub_date
+			now = timezone.now()
+			while now>=oldest_blog_date or (now.month==oldest_blog_date.month and now.year==oldest_blog_date.year):
+				context['dates'].append((now.strftime('%B'),now.year))
+				now = now - relativedelta(months=1)
 		return render(request, 'blogger/blog.html', context)
 
 class BlogPostView(View):
@@ -104,6 +151,7 @@ class BlogPostView(View):
 		return render(request, 'blogger/blogpost.html', context)
 
 @method_decorator(login_required, name='get')
+@method_decorator(login_required, name='post')
 class BlogManagerView(View):
 	def get(self, request):
 		blog = get_object_or_404(Blog, user=request.user)
@@ -111,8 +159,53 @@ class BlogManagerView(View):
 		context = {
 			'username': request.user.username,
 			'blog': blog,
-			'blogposts': blogposts
+			'blogposts': [],
+			'dates': []
 		}
+		for blogpost in blogposts:
+			image_count = blogpost.imagedoc_set.all().count()
+			if image_count>0:
+				context['blogposts'].append((blogpost,blogpost.imagedoc_set.all()[image_count-1], blogpost.blogpost_text[:PREVIEW_TEXT_CHAR]))
+			else:
+				context['blogposts'].append((blogpost, '', blogpost.blogpost_text[:PREVIEW_TEXT_CHAR]))
+		if blogposts.count()>0:
+			oldest_blog_date = blogposts[blogposts.count()-1].pub_date
+			now = timezone.now()
+			while now>=oldest_blog_date or (now.month==oldest_blog_date.month and now.year==oldest_blog_date.year):
+				context['dates'].append((now.strftime('%B'),now.year))
+
+				now = now - relativedelta(months=1)
+		return render(request, 'blogger/blogmanager.html', context)
+	def post(self, request):
+		blog = get_object_or_404(Blog, user=request.user)
+		context = {
+			'username': request.user.username,
+			'blog': blog,
+			'blogposts': [],
+			'dates': [],
+		}
+		if 'datefilter' in request.POST:
+			datetime = request.POST['datefilter']
+			if datetime == 'All Time':
+				blogposts = blog.blogpost_set.all().order_by('-pub_date')
+			else:
+				datetime = datetime.split(' ')
+				month = convertMonthNameToNumber(datetime[0])
+				year = datetime[1]
+				blogposts = blog.blogpost_set.all().order_by('-pub_date').filter(pub_date__month=month, pub_date__year=year)
+		for blogpost in blogposts:
+			image_count = blogpost.imagedoc_set.all().count()
+			if image_count>0:
+				context['blogposts'].append((blogpost,blogpost.imagedoc_set.all()[image_count-1], blogpost.blogpost_text[:PREVIEW_TEXT_CHAR]))
+			else:
+				context['blogposts'].append((blogpost, '', blogpost.blogpost_text[:PREVIEW_TEXT_CHAR]))
+		blogposts = blog.blogpost_set.all().order_by('-pub_date')
+		if blogposts.count()>0:
+			oldest_blog_date = blogposts[blogposts.count()-1].pub_date
+			now = timezone.now()
+			while now>=oldest_blog_date or (now.month==oldest_blog_date.month and now.year==oldest_blog_date.year):
+				context['dates'].append((now.strftime('%B'),now.year))
+				now = now - relativedelta(months=1)
 		return render(request, 'blogger/blogmanager.html', context)
 
 @method_decorator(login_required, name='get')
@@ -120,18 +213,35 @@ class BlogManagerView(View):
 class AccountManagerView(View):
 	def get(self, request):
 		blog = get_object_or_404(Blog, user=request.user)
+		form_content = {
+			'blog_description': blog.blog_description,
+			'blog_background_color': blog.blog_background_color,
+		}
+		form = BlogEditForm(form_content)
 		context = {
 			'username': request.user.username,
 			'blog': blog,
+			'form': form,
 		}
 		return render(request, 'blogger/accountmanager.html', context)
 	def post(self, request):
 		user = request.user
-		if 'delete' in request.POST:
-			logout(request)
-			user.delete()
-			return render(request, 'blogger/index.html')
-		return render(request, 'blogger/accountmanager.html')
+		form = BlogEditForm(request.POST, request.FILES)
+		blog = get_object_or_404(Blog, user=request.user)
+		if form.is_valid():
+			if 'delete' in request.POST:
+				logout(request)
+				user.delete()
+				return HttpResponseRedirect(reverse('blogger:index'))
+				# return render(request, 'blogger/index.html')
+			
+			blog.blog_description = form.cleaned_data['blog_description']
+			blog.blog_background_color = form.cleaned_data['blog_background_color']
+			if request.FILES.get('blog_banner'):
+				blog.blog_banner = form.cleaned_data['blog_banner']
+			blog.save()
+		return HttpResponseRedirect(reverse('blogger:blogmanager'))
+		# return render(request, 'blogger/accountmanager.html')
 
 @method_decorator(login_required, name='get')
 @method_decorator(login_required, name='post')
@@ -243,3 +353,31 @@ def addBlogPostComment(request, blog_title, blogpost_id, blogpost_title):
 		return redirect_to
 	else:
 		return redirect_to
+
+def convertMonthNameToNumber(month):
+	if month == 'January':
+		return 1
+	elif month == 'February':
+		return 2
+	elif month == 'March':
+		return 3
+	elif month == 'April':
+		return 4
+	elif month == 'May':
+		return 5
+	elif month == 'June':
+		return 6
+	elif month == 'July':
+		return 7
+	elif month == 'August':
+		return 8
+	elif month == 'September':
+		return 9
+	elif month == 'October':
+		return 10
+	elif month == 'November':
+		return 11
+	elif month == 'December':
+		return 12
+	else:
+		return 0
